DicomText - move string utilities from DicomText.py into StringUtils.py
